+++ src/main/java/twilightforest/entity/boss/EntityTFYetiAlpha.java
@@ -305,10 +305,10 @@ public class EntityTFYetiAlpha extends MonsterEntity implements IRangedAttackMob
 
 			// [VanillaCopy] Part of EntitySkeleton.attackEntityWithRangedAttack
 			double d0 = target.getPosX() - this.getPosX();
-			double d1 = target.getBoundingBox().minY + (double) (target.getHeight() / 3.0F) - ice.getPosY();
+			double d1 = target.getBoundingBox().minY + target.getHeight() / 3.0F - ice.getPosY();
 			double d2 = target.getPosZ() - this.getPosZ();
-			double d3 = (double) MathHelper.sqrt(d0 * d0 + d2 * d2);
-			ice.shoot(d0, d1 + d3 * 0.20000000298023224D, d2, 1.6F, (float) (14 - this.world.getDifficulty().getId() * 4));
+			double d3 = MathHelper.sqrt(d0 * d0 + d2 * d2);
+			ice.shoot(d0, d1 + d3 * 0.20000000298023224D, d2, 1.6F, 14 - this.world.getDifficulty().getId() * 4);
 
 			this.playSound(SoundEvents.ENTITY_ARROW_SHOOT, 1.0F, 1.0F / (this.getRNG().nextFloat() * 0.4F + 0.8F));
 			this.world.addEntity(ice);
